Other website is scraped
The old website had multiple entries for Pokemon from older versions. This was a
pain to deal with so I switched to another site only representing Galar Pokemon.
This may lead to a small error (e.g. I only consider Galar-Mauzi, but not the
standard Mauzi, although both can be obtained in the game).